flashtext2 (cython)/pyproject.toml
[tool.poetry]
name = "flashtext-cy"
version = "0.1.0"
description = ""
authors = ["shner-elmo <770elmo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"

[tool.poetry.group.dev.dependencies]
Cython = "^0.29.34"
ipython = "^8.12.0"
flashtext = "^2.7"
requests = "^2.28.2"
regex = "^2023.3.23"
pandas = "^2.0.0"
matplotlib = "^3.7.1"
ez-stopwatch = "^1.0.1"
py-spy = "^0.3.14"
flashtext2 = "^0.1.0"
more-itertools = "^9.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
